+++ packages/vue/src/components/index.ts
@@ -39,6 +39,7 @@ export { default as TSelect } from './TSelect.vue';
 export type { TSelectOption } from './TSelect.vue';
 export { default as TSidebar } from './TSidebar.vue';
 export type { TSidebarSide } from './TSidebar.vue';
+export { default as TAppShell } from './TAppShell.vue';
 export { default as TInput } from './TInput.vue';
 export { default as TNumberInput } from './TNumberInput.vue';
 export { default as TSelectableList } from './TSelectableList.vue';