+++ apps/landing/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @treeui/landing
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @treeui/vue@0.12.0
+
 ## 0.1.3
 
 ### Patch Changes